main_event/requirements.txt
Flask==3.0.0
qrcode==7.4.2
Pillow==10.1.0
python-dateutil==2.8.2